docs(naming) : rename model to lowercase
